app/requirements.txt
Flask==1.1.2
requests==2.24.0
google-api-python-client==1.12.8
python-dotenv==0.15.0